Fairly big changes to the walker
The first change allows the walker to offer up the information needed by
movers without the mover needing friend access, this allows easier adding of
movers by the user, or by me later

The second change, in the vein of easier addition of mover classes later is
to change the walker such that no decision making happens in the walker. All
decision about accepting the proposal the mover generated happens in the
mover. In this way, movers with significantly acceptance algorithms can
easily be handled. All the mover needs to do is call the jumpToNewPoint or
stayAtCurrentPoint functions for the appropriate things to be done to the
chain.

Signed-off-by: James Till Matta <jamesmatta@gmail.com>